requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-barcode==0.14.0
Pillow==11.1.0
Flask==2.3.3
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
certifi>=2024.2.2 